refactor(registry): move lockfile tests behind facade
